Move testing modules into a separate sub-module
For better encapsulation.
